Moved Post/index.tsx -> Posts/Post.tsx & fixed TS error
